// frontend/src/types/user.ts
// Интерфейс `Role` определяет структуру объекта роли пользователя.
// Это помогает типизировать, какие значения может принимать поле 'role' у пользователя.
export interface Role {
    id: number; // Уникальный идентификатор роли (например, 1 для админа, 2 для сотрудника).
    // Поле `role` строго типизировано, чтобы принимать только одно из трех предопределенных строковых значений.
    role: 'admin' | 'staff' | 'client'; // Строковое представление роли пользователя.
}

// Интерфейс `User` определяет структуру данных пользователя, как они могут быть получены с сервера
// или храниться в состоянии приложения.
export interface User {
    id: number; // Уникальный идентификатор пользователя.
    login_client: string; // Логин пользователя.
    phone_number_client: string; // Номер телефона клиента.
    company_name: string; // Название компании, связанной с клиентом.
    role: Role; // Объект роли пользователя, типизированный интерфейсом `Role`.
    [key: string]: any; // Индексная сигнатура: позволяет добавлять любые другие свойства к объекту `User`
                        // без явного объявления, что может быть полезно, если структура данных
                        // пользователя может варьироваться или расширяться.
}

// Интерфейс `AuthState` определяет структуру состояния аутентификации в Redux-хранилище.
// Он содержит всю необходимую информацию о статусе аутентификации пользователя.
export interface AuthState {
    token: string | null; // Токен аутентификации (JWT), может быть строкой или null, если пользователь не аутентифицирован.
    user: User | null; // Данные текущего аутентифицированного пользователя, типизированные интерфейсом `User`, или null.
    isAuthenticated: boolean; // Флаг, указывающий, аутентифицирован ли пользователь.
    loading: boolean; // Флаг, указывающий, идет ли процесс аутентификации или загрузки данных пользователя.
    error: string | null; // Сообщение об ошибке, если аутентификация не удалась, или null.
}